modules: split membership
